src/co/uk/vonaq/cableguardian/utils/RegexUtils.java
package co.uk.vonaq.cableguardian.utils;

import java.util.regex.Matcher;
import java.util.regex.Pattern;

public class RegexUtils {

    private static final String EMAIL_PATTERN = "^[_A-Za-z0-9-\\+]+(\\.[_A-Za-z0-9-]+)*@[A-Za-z0-9-]+(\\.[A-Za-z0-9]+)*(\\.[A-Za-z]{2,})$";

    public static boolean validEmail(final String hex) {
        Pattern pattern = Pattern.compile(EMAIL_PATTERN);
        Matcher matcher = pattern.matcher(hex);
        return matcher.matches();
    }

    public static boolean areNumbersSame(String first, String second) {
        System.out.println("areNumbersSame "+first+" "+second);
        return (first.trim().endsWith(second.substring(second.length() - 5)));
    }
}
